+++ comps/login/AccountCreate.js
@@ -7,7 +7,7 @@ import {
   TouchableOpacity,
   TouchableHighlight
 } from 'react-native';
-import styles from '../styles/DonatePen';
+import styles from '../../styles/DonatePen';
 import {Actions} from 'react-native-router-flux'
 import axios from 'axios';
 
@@ -37,7 +37,7 @@ function AccountCreate({text}) {
   return (
     <View>
       <View style={styles.mainbody}>
-        <Image style={styles.imgPosted} source={require("../assets/img/created_account.png")} />
+        <Image style={styles.imgPosted} source={require("../../assets/img/created_account.png")} />
 
         <View style={styles.headers}>
           <Text style={styles.titleHeader}>Your account has been created!</Text>
